etl/pyproject.toml
[tool.poetry]
name = "etl"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
psycopg = "^3.1.8"
python-dotenv = "^0.21.1"
elasticsearch = "^8.6.2"
dateutils = "^0.6.12"
redis = "^4.5.1"
pydantic = "^1.10.5"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
